refactor: 패키지 변경 및 이름 변경
