+++ data/src/main/java/com/wakeup/data/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.wakeup.data.di
 
 import com.wakeup.data.repository.PlaceSearchRepositoryImpl
-import com.wakeup.data.repository.moment.MomentRepositoryImpl
+import com.wakeup.data.repository.MomentRepositoryImpl
 import com.wakeup.domain.repository.MomentRepository
 import com.wakeup.domain.repository.PlaceSearchRepository
 import dagger.Binds
