+++ data/src/main/java/com/wakeup/data/repository/PlaceSearchRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.wakeup.data.repository
 
 import com.wakeup.data.network.response.toDomain
-import com.wakeup.data.source.remote.PlaceSearch.PlaceSearchRemoteDataSource
+import com.wakeup.data.source.remote.placeSearch.PlaceSearchRemoteDataSource
 import com.wakeup.domain.model.Place
 import com.wakeup.domain.repository.PlaceSearchRepository
 import javax.inject.Inject
